requirements.txt
python-dotenv
openai
nltk